Passthrough endpoints added
The passthrough endpoints that work are:
- /forecast for NWS forecasts
- /adds for ADDS data
- /usno for USNO data

A note on the USNO endpoint: it would seem that the Navy either signed
their own SSL certificate or else they just don't have SSL set up
correctly, so I disabled CA checking when fetching from this endpoint.
It seems like an SSL endpoint for this is relatively new, so this might
just be a deployment problem that will be fixed soon. I'll check on this
later.